requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alchemy-sdk==0.1.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
bitarray==3.6.0
certifi==2024.8.30
charset-normalizer==3.4.1
ckzg==2.1.1
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
cytoolz==1.0.1
dataclass-wizard==0.35.1
Django==5.1.5
docutils==0.21.2
et-xmlfile==1.1.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
fastapi==0.115.5
fonttools==4.55.0
frozenlist==1.7.0
hexbytes==1.3.1
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jsonschema==4.25.0
joblib==1.4.2
kafka-python==2.2.15
keyring==25.6.0
kiwisolver==1.4.7
markdown-it-py==3.0.0
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.6.3
nh3==0.2.20
nltk==3.9.1
numpy==2.1.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
pillow==11.0.0
pkginfo==1.12.0
propcache==0.3.2
psycopg2==2.9.10
pycryptodome==3.23.0
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
pyunormalize==16.0.0
readme_renderer==44.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rlp==4.1.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.8.0
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.14.2
sqlparse==0.5.3
starlette==0.41.3
threadpoolctl==3.5.0
toolz==1.0.0
tqdm==4.67.0
twine==6.0.1
types-requests==2.32.4.20250611
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.35.0
web3==7.13.0
websockets==15.0.1
wheel==0.45.1
xlrd==2.0.1
yarl==1.20.1
